Added level selection highlight (+147B).

+++ src/engine/sys.h
@@ -186,7 +186,7 @@ namespace Sys {
 			};
 
 			u16			box_min[2];
-			u8			box_size[2];
+			u16			box_size[2];
 			i8			anchor[2];
 			u8			advance;
 		};
